HoS Samurai
Long ago in a distant sector, I, Nuclear Operative, the chameleon clothing nukie of darkness unleashed an unspeakable war declaration, but a foolish Head of Security warrior wielding an Admeme sword stepped forth to oppose me!

Before the final blow was struck, I throw open a maxcap in time and flung him into SS14 where my evil is strong!
Now the fool seeks to return to the past and undo the future that is NUKIE MAJOR!!!