Melee Part 6: Stamina Issues and AttackRates (#2319)
This PR is a variety of touch ups to many weapons to help make them more
unique, a little more powerful, and also to fix a few issues.
Some weapons still had a stamina cost of 5 or 10, this is almost
universally lowered to a new default of 0.5 or 2.5, with a few
exceptions.
The generic power attack angle is now increased as well which should
include combat feel.

Stamina damage was also lowered on a number of weapons.

The 'issue' of some melee weapons ended up being that their attack rates
were going off of Wizden numbers, which is wrong. 0.75 would be a slower
swing on a melee in Wizden code, but here it is 0.75s(when 1s is the
default). A variety of melee weapons had this issue where either their
left click or their power attack

---
<details><summary><h1>Media</h1></summary>
Showcasing the difference of a bugged power attack with the wrench vs
it's normal left click
<p>

![rotating](https://github.com/user-attachments/assets/330828c9-72b4-495e-a772-c4c2b84daa4a)

</p>
</details>

---

:cl: ODJ
- add: Changeling now has power attacks for it's blade and hammer.
- tweak: Melee weapons given slight adjustments to feel better and more
unique.

---------

Co-authored-by: jcsmithing <jcsmithing@gmail.com>
Co-authored-by: Timfa <timfalken@hotmail.com>