Morgan gives me Backshots in Aghost, send help
Hello hello helloooo~! Morgan yelled at me to "Drop my pants and touch my toes, she's gonna show me where the wild kibby goes!". I'm not sure what she means by this, but I believe in the the heart of the cards. It's not fate, it's destiny *wild nickleback-esque guitar riffs*.

Anyways, tonight on Gregg Gear: Gregg gets his head stuck in a fence and needs to be fully soaked in butter to be wriggled free, Fox duels a Embden Goose over the usage of the local bus systems and KC engages in a pie eating contest against ghost of that weaselly traitor fuck Stonewall Jackson.

Mono receives other worldly visions after having a temperature of 120°F (it's okay, Mono and the Moose from last episode have entered an uneasy alliance and the Moose is caring for him), TJ goes on an epic journey to save Valiant from the clutches of the Taliban and June is stranded on the Moon with nothing but her big, bug ass to keep her company.

Jenn flys to Vietnam and experiences a culinary master-tour of such a wonderful land, April gets trapped in VR matrix-style and has to beat the final boss of VR, Gabe Newell. Finally, Cami gets tied to a chair and told how wonderful and important she is to everyone.