Papa gets silly when he has the Moscato at Noon
Oi oi oi! In honor of Anzac Day, we're here to remember the loss of our beautiful baby boy TJohnson to the war against shitters in the Delta-V Sector. May his handsome bones rest in eternal, cigarette filled slumber.

Anyways, tonight on Gregg Gear: Gregg becomes a desiccated husk after being tossed into a pit of ravenous wolf-girls, Mono and the Moose have now entered the third act of their Enemies to Lovers Slowburn arc, and TJ has rescued Valiant from the Taliban after a deluge of slapstick comedy that would be found in an Ace Ventura movie and hyper-violent urban assault tactics.

KC braves the depths of Lethe while leading some of the bravest Marines in the Corps, June joins the UNMC after watching the Alien movies (this will become important in the sequel) and Jenn becomes a local crime mob boss and runs a night club (it will actually be very nice and comfortable, a sanctuary for wayward souls in the night).

April and Kira are arrested for toxic Yuri (multiple counts of homicide), Morgan is turned into a large Mallard Duck after crossing a wizard and failing to answer his riddles three and Fox enters a bake sale without reading the fine print and soon becomes entrapped in a bake off... to the DEATH.

Cami receives a healthy mix of affection and beatings for their wonderful new hit album they released and Erin alongside Graves enter a new and magical world (New Brunswick) and become enraptured in the local culture (theft).