Arena/Asterisk/Edge Tweaks
Adding a Document Printer and Vendomat at the Tool Room in Asterisk and two vending machines in Security's Lobby

Added a Document Printer in Arena

Updated Edge's Dorms to be more enjoyable and protected from Meteors by request of April. Also updated the stripper room to make it better and more private.
Updated Security, to make the HoS's bedroom seperate from their office.
Same with the Warden.
Shrank the MASSIVE Corpsman's Office to do so
Overhauled the Security Ready Room, it's break room, it's Armory and added a Document Printer to the Front Desk
 area.
Also added a Document printer to the Library here on Edge.